move test helper functions back
